+++ picklepipe/picklepipe.py
@@ -29,7 +29,7 @@ class PicklePipe(BaseSerializingPipe):
     socket as a interface to send pickled objects to a peer.
     Can be used to pickle not only single objects but also
     to pickle objects in a stream-able fashion. """
-    def __init__(self, sock, protocol=None):
+    def __init__(self, sock, protocol=None, max_size=None):
         """
         Creates a :class:`picklepipe.PicklePipe` instance wrapping
         a given socket.
@@ -37,7 +37,7 @@ class PicklePipe(BaseSerializingPipe):
         :param sock: Socket to wrap.
         :param protocol: Pickling protocol to favor.
         """
-        super(PicklePipe, self).__init__(sock, None)
+        super(PicklePipe, self).__init__(sock, None, max_size=max_size)
         self._protocol = protocol
         self._protocol_sent = False
         self._protocol_recv = False